update to use ms sql

+++ InColUn/src/InColUn.Db/Models/User.cs
@@ -4,10 +4,10 @@ namespace InColUn.Db.Models
 {
     public class User
     {
-        public ulong Id { get; set; }
+        public long Id { get; set; }
         public string login_string { get; set; }
         public string password_hash { get; set; }
-        public int? salt { get; set; }
+        public long? salt { get; set; }
         public string display_name { get; set; }
         public string email { get; set; }
         public string auth_provider { get; set; }
